clarify specification compliance (#35)
There's nothing in tl right now that documents the spec compliance status (whether it tries to follow it or not, see #34). 
This PR adds some information to the README regarding this, and changes the benchmark section. It should be a lot clearer which parsers attempt to comply with the specification and which ones don't, as well as making it clear that it's more of a theoretical benchmark and doesn't necessarily say something meaningful (at least comparing tl to html5ever and lol-html), since the performance of a parser that attempts to follow the specification can't be compared to one that doesn't (due to the fact that the spec really limits what one can do).